+++ Dev/Source/Engine/Source/FreStringUtils.cpp
@@ -6,7 +6,7 @@ namespace FRE
 {
 	namespace Utils
 	{
-		int SPrintf_s(char * buffer, size_t size, const char * format, ...)
+		int SPrintf_s(char * buffer, unsigned size, const char * format, ...)
 		{
 			if (!size)
 				return 0;
